requirements.txt
absl-py==0.5.0
astor==0.7.1
dj-database-url==0.5.0
Django==2.1.1
django-braces==1.13.0
django-heroku==0.3.1
djangorestframework==3.8.2
gast==0.2.0
grpcio==1.15.0
gunicorn==19.9.0
h5py==2.8.0
Keras==2.2.2
Keras-Applications==1.0.5
Keras-Preprocessing==1.0.2
Markdown==3.0.1
numpy==1.14.5
protobuf==3.6.1
psycopg2==2.7.5
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
Werkzeug==0.14.1
whitenoise==4.0
